validations on Text and Translation are now in Repository class in order to be reused everywhere
