+++ OpenTranslator/Repostitory/TranslationRepository.cs
@@ -1,7 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 
 using OpenTranslator.Data;
+using OpenTranslator.Models.Input;
 
 namespace OpenTranslator.Repostitory
 {
@@ -39,7 +41,7 @@ namespace OpenTranslator.Repostitory
         /// Returns a list of all Text objects in db
         /// </summary>
         /// <returns></returns>
-		public IEnumerable<Text> GetText()
+		public IEnumerable<Text> GetText() 
 		{
 			return GetDbContext().Texts.ToList();
 		}
@@ -75,6 +77,38 @@ namespace OpenTranslator.Repostitory
 			return GetDbContext().Translations.Where(x=>x.TextId==TextId).ToList();
 		}
 
+        /// <summary>
+        /// Check if Translation entity already exists in storage DB
+        /// </summary>
+        /// <param name="input"></param>
+        /// <param name="translatedText"></param>
+        /// <returns></returns>
+        public bool IsTranslateAlreadyStoraged(AdminInput input, string translatedText)
+        {
+            var textQuery = GetAll().Where(x => x.TextId == input.TextId && x.LanguageCode == input.LanguageCode && x.Translated_Text.Equals(translatedText));
+
+            return textQuery.FirstOrDefault() != null;
+        }
+
+        /// <summary>
+        /// Check if Text entity is already storaged in DB
+        /// </summary>
+        /// <param name="input"></param>
+        /// <returns></returns>
+        public bool IsTextIdAlreadyStoraged(AdminInput input)
+        {
+            var textQuery = GetText().Where(x => x.TextId == input.TextId);
+
+            if (input.Id != 0)
+            {
+                textQuery = textQuery.Where(y => y.Id != Convert.ToDecimal(input.Id));
+            }
+
+            var text = textQuery.FirstOrDefault();
+
+            return text != null;
+        }
+
         #endregion
-	}
+    }
 }
